Rename the module to cryptography.cobblestone, trim docs
Per review: the module, docs page, stubs, and Rust module are renamed
from chunked_encryption to cobblestone (the classes were already named
for the Cobblestone instantiations), and the docs' Implementation
section is dropped since the page already states it implements the
C2SP chunked-encryption specification.
