+++ src/cryptography/hazmat/bindings/_rust/cobblestone.pyi
