+++ src/rust/src/lib.rs
@@ -34,7 +34,7 @@ use crate::error::CryptographyResult;
 mod asn1;
 mod backend;
 mod buf;
-mod chunked_encryption;
+mod cobblestone;
 mod declarative_asn1;
 mod error;
 mod exceptions;
@@ -133,7 +133,7 @@ mod _rust {
     #[pymodule_export]
     use crate::asn1::asn1_mod;
     #[pymodule_export]
-    use crate::chunked_encryption::chunked_encryption_mod;
+    use crate::cobblestone::cobblestone_mod;
     #[pymodule_export]
     use crate::exceptions::exceptions;
     #[pymodule_export]
